Replace handlers to correct template
